KeyEvent: consolidate conversions and unify types
- Replace 9 hand-rolled copies with single conversion call
- Unify to uint32_t, fix Android nanosecond bug

+++ dali/internal/system/tizen/widget-application-impl-tizen.cpp
@@ -87,7 +87,7 @@ bool OnKeyEventCallback(const char* id, screen_connector_event_type_e eventType,
   std::string        keyEventName = keyName ? std::string(keyName) : std::string();
   const Dali::String emptyString("");
   Dali::KeyEvent     event = Dali::DevelKeyEvent::New(Dali::String(keyEventName.c_str()), emptyString, emptyString, keyCode, 0,
-                                                      static_cast<unsigned long>(timestamp),
+                                                      static_cast<uint32_t>(timestamp),
                                                       state, emptyString, emptyString,
                                                       Device::Class::NONE, Device::Subclass::NONE);
